search_system/requirements.txt
flask>=3.0.0
openai>=1.0.0
numpy>=1.24.0
scipy>=1.11.0
pypdf>=3.0.0
python-docx>=1.0.0
